#36473528 Корректировка логики платного создания вакансии.

+++ LeokaEstetica.Platform.Processing/BuilderData/OrderData.cs
@@ -29,7 +29,7 @@ internal class OrderData
     /// <summary>
     /// Id тарифа.
     /// </summary>
-    public long? RuleId { get; set; }
+    public int? RuleId { get; set; }
 
     /// <summary>
     /// Тип заказа.
